gulpfile split into several smaller JS files
The tasks now live inside gulp>tasks
