+++ gulp/tasks/watch.js
@@ -1,0 +1,27 @@
+var gulp = require('gulp'),
+    browserSync = require('browser-sync').create();
+
+
+
+
+
+// Main Watch
+gulp.task('watch', function() {
+  // TO DO: dest has to be mandatory or have a deault folder.
+  var argv = require('yargs')
+    .usage('\nError.\nYou have to specify the folder of the unit you want Gulp to watch.\nE.g: gulp -f 300x250')
+    .demand(['f'])
+    .argv;
+
+  var dest = 'build/' + argv.f;
+  console.log("Watching folder: ", argv.f, dest);
+  browserSync.init({
+    server: dest
+  });
+
+  // TO DO: This does not seem to be watching all image folders
+  gulp.watch('src/**/*.{gif,jpg,png,svg}', ['watch-images']);
+  gulp.watch('src/**/*.css', ['build-css']);
+  gulp.watch('src/**/*.html', ['watch-html']);
+  gulp.watch('src/**/*.js', ['watch-js']);
+});
