+++ package.json
@@ -18,6 +18,8 @@
     "gulp-uglify": "^1.5.1",
     "gulp-zip": "^3.0.2",
     "path": "^0.11.14",
+    "require-dir": "^0.3.0",
+    "run-sequence": "^1.1.5",
     "yargs": "^3.31.0"
   },
   "scripts": {
