+++ src/shared/js/Base.js
@@ -8,8 +8,8 @@ var Ad = function() {
 
   // Get the ad's size for future use
   var stage = {
-    w:document.getElementById("ad").offsetWidth,
-    h:document.getElementById("ad").offsetHeight
+    w:document.querySelector("ad-banner").offsetWidth,
+    h:document.querySelector("ad-banner").offsetHeight
   };
 
   // Instantiate the necessary constructors
@@ -52,7 +52,7 @@ var Ad = function() {
   // TO DO: make a bg image to use
   // ldr.load("ad-banner", ["bg"], ".jpg");
   // Or you can target specific id names
-  ldr.load("svgText", []"here, is, a, bunch, of, text"], ".svg");
+  // ldr.load("svgText", ["here, is, a, bunch, of, text"], ".svg");
 
 
   // add the buttons
